Move the user initialization code in config/initializers/ideals.rb to db/seeds.rb
This code doesn't belong in an initializer because it touches the database; remember that initializers may be run before a database exists, such as on db:create. It doesn't work well in a seed either, so there is a TODO to address that. Just trying to clear a path for a Rails 6 upgrade.